# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "College Hierarchy API",
    "version": "1.0.0",
    "description": "API documentation for the College Hierarchy system"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Server"
    }
  ],
  "paths": {
...
    "/professor/addprofessor": {
      "post": {
        "summary": "Add a new Professor",
        "description": "Creates a new Professor in the college hierarchy.",
        "operationId": "addProfessor",
        "tags": ["Professor"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Dr. Bhanu" },
                  "parentId": { "type": "integer", "example": 1 }
                },
                "required": ["name", "parentId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Professor added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Professor added successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 2 },
                        "name": { "type": "string", "example": "Dr. Bhanu" },
                        "hodId": { "type": "integer", "example": 1 }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Name and Parent ID are required" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Cannot add Professor" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/professor/getprofessor": {
      "get": {
        "summary": "Get all Professors",
        "description": "Retrieves all Professors.",
        "operationId": "getProfessor",
        "tags": ["Professor"],
        "responses": {
          "200": {
            "description": "Professors fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer", "example": 2 },
                      "name": { "type": "string", "example": "Dr. Bhanu" },
                      "hodId": { "type": "integer", "example": 1 }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Server Error" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/professor/getprofessorbyid/{id}": {
      "get": {
        "summary": "Get Professor by ID",
        "description": "Retrieves a Professor by ID.",
        "operationId": "getProfessorById",
        "tags": ["Professor"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 2 }
          }
        ],
        "responses": {
          "200": {
            "description": "Professor fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 2 },
                        "name": { "type": "string", "example": "Dr. Bhanu" },
                        "hodId": { "type": "integer", "example": 1 }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Invalid ID format" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Professor not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Professor with 2 not found (controller)" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Server Error" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/professor/getallbyid/{id}": {
      "get": {
        "summary": "Get Professor and Child Data by ID",
        "description": "Retrieves a Professor along with associated Students by Professor ID.",
        "operationId": "getAllProfById",
        "tags": ["Professor"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 2 }
          }
        ],
        "responses": {
          "200": {
            "description": "Professor and child data fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Professor data fetched successfully" },
                    "data": { "type": "object" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Id is required" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Professor not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Professor not found or delete failed" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Server Error" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/professor/deleteprofessor/{id}": {
      "delete": {
        "summary": "Delete Professor",
        "description": "Deletes a Professor and related child data.",
        "operationId": "deleteProfessor",
        "tags": ["Professor"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 2 }
          }
        ],
        "responses": {
          "200": {
            "description": "Professor deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Professor ID 2 deleted successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Id is required" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Professor not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "id not found or delete failed" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Cannot delete Professor" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/student/deletestudent/{id}": {
      "delete": {
        "summary": "Delete Student",
        "description": "Deletes a Student.",
        "operationId": "deleteStudent",
        "tags": ["Student"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Student id: 1 deleted successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Id is required" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "id not found or delete failed" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Cannot delete Student" },
                    "error": { "type": "string", "example": "Database error" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}